refactor: move PlateauGSITileOption class to its own file for better organization
